requirements.txt
requests == 2.24.0
paho-mqtt == 1.5.1
jsonify == 0.5